LayoutTests/fast/dom/constants-expected.txt

This page tests CSSRule, CSSValue, NodeFilter, and Event. It tests:

Whether their global constructors have the correct constant values
Whether their objects have the correct constant values
PASS: cssRule.UNKNOWN_RULE should be 0 and is.
PASS: cssRule.STYLE_RULE should be 1 and is.
PASS: cssRule.CHARSET_RULE should be 2 and is.
PASS: cssRule.IMPORT_RULE should be 3 and is.
PASS: cssRule.MEDIA_RULE should be 4 and is.
PASS: cssRule.FONT_FACE_RULE should be 5 and is.
PASS: cssRule.PAGE_RULE should be 6 and is.
PASS: window.CSSRule.UNKNOWN_RULE should be 0 and is.
PASS: window.CSSRule.STYLE_RULE should be 1 and is.
PASS: window.CSSRule.CHARSET_RULE should be 2 and is.
PASS: window.CSSRule.IMPORT_RULE should be 3 and is.
PASS: window.CSSRule.MEDIA_RULE should be 4 and is.
PASS: window.CSSRule.FONT_FACE_RULE should be 5 and is.
PASS: window.CSSRule.PAGE_RULE should be 6 and is.
PASS: cssValue.CSS_INHERIT should be 0 and is.
PASS: cssValue.CSS_PRIMITIVE_VALUE should be 1 and is.
PASS: cssValue.CSS_VALUE_LIST should be 2 and is.
PASS: cssValue.CSS_CUSTOM should be 3 and is.
PASS: window.CSSValue.CSS_INHERIT should be 0 and is.
PASS: window.CSSValue.CSS_PRIMITIVE_VALUE should be 1 and is.
PASS: window.CSSValue.CSS_VALUE_LIST should be 2 and is.
PASS: window.CSSValue.CSS_CUSTOM should be 3 and is.
PASS: nodeFilter.FILTER_ACCEPT should be 1 and is.
PASS: nodeFilter.FILTER_REJECT should be 2 and is.
PASS: nodeFilter.FILTER_SKIP should be 3 and is.
PASS: nodeFilter.SHOW_ALL should be 4294967295 and is.
PASS: nodeFilter.SHOW_ELEMENT should be 1 and is.
PASS: nodeFilter.SHOW_ATTRIBUTE should be 2 and is.
PASS: nodeFilter.SHOW_TEXT should be 4 and is.
PASS: nodeFilter.SHOW_CDATA_SECTION should be 8 and is.
PASS: nodeFilter.SHOW_ENTITY_REFERENCE should be 16 and is.
PASS: nodeFilter.SHOW_ENTITY should be 32 and is.
PASS: nodeFilter.SHOW_PROCESSING_INSTRUCTION should be 64 and is.
PASS: nodeFilter.SHOW_COMMENT should be 128 and is.
PASS: nodeFilter.SHOW_DOCUMENT should be 256 and is.
PASS: nodeFilter.SHOW_DOCUMENT_TYPE should be 512 and is.
PASS: nodeFilter.SHOW_DOCUMENT_FRAGMENT should be 1024 and is.
PASS: nodeFilter.SHOW_NOTATION should be 2048 and is.
PASS: window.NodeFilter.FILTER_ACCEPT should be 1 and is.
PASS: window.NodeFilter.FILTER_REJECT should be 2 and is.
PASS: window.NodeFilter.FILTER_SKIP should be 3 and is.
PASS: window.NodeFilter.SHOW_ALL should be 4294967295 and is.
PASS: window.NodeFilter.SHOW_ELEMENT should be 1 and is.
PASS: window.NodeFilter.SHOW_ATTRIBUTE should be 2 and is.
PASS: window.NodeFilter.SHOW_TEXT should be 4 and is.
PASS: window.NodeFilter.SHOW_CDATA_SECTION should be 8 and is.
PASS: window.NodeFilter.SHOW_ENTITY_REFERENCE should be 16 and is.
PASS: window.NodeFilter.SHOW_ENTITY should be 32 and is.
PASS: window.NodeFilter.SHOW_PROCESSING_INSTRUCTION should be 64 and is.
PASS: window.NodeFilter.SHOW_COMMENT should be 128 and is.
PASS: window.NodeFilter.SHOW_DOCUMENT should be 256 and is.
PASS: window.NodeFilter.SHOW_DOCUMENT_TYPE should be 512 and is.
PASS: window.NodeFilter.SHOW_DOCUMENT_FRAGMENT should be 1024 and is.
PASS: window.NodeFilter.SHOW_NOTATION should be 2048 and is.
PASS: event.NONE should be 0 and is.
PASS: event.CAPTURING_PHASE should be 1 and is.
PASS: event.AT_TARGET should be 2 and is.
PASS: event.BUBBLING_PHASE should be 3 and is.
PASS: event.MOUSEDOWN should be 1 and is.
PASS: event.MOUSEUP should be 2 and is.
PASS: event.MOUSEOVER should be 4 and is.
PASS: event.MOUSEOUT should be 8 and is.
PASS: event.MOUSEMOVE should be 16 and is.
PASS: event.MOUSEDRAG should be 32 and is.
PASS: event.CLICK should be 64 and is.
PASS: event.DBLCLICK should be 128 and is.
PASS: event.KEYDOWN should be 256 and is.
PASS: event.KEYUP should be 512 and is.
PASS: event.KEYPRESS should be 1024 and is.
PASS: event.DRAGDROP should be 2048 and is.
PASS: event.FOCUS should be 4096 and is.
PASS: event.BLUR should be 8192 and is.
PASS: event.SELECT should be 16384 and is.
PASS: event.CHANGE should be 32768 and is.
PASS: window.Event.NONE should be 0 and is.
PASS: window.Event.CAPTURING_PHASE should be 1 and is.
PASS: window.Event.AT_TARGET should be 2 and is.
PASS: window.Event.BUBBLING_PHASE should be 3 and is.
PASS: window.Event.MOUSEDOWN should be 1 and is.
PASS: window.Event.MOUSEUP should be 2 and is.
PASS: window.Event.MOUSEOVER should be 4 and is.
PASS: window.Event.MOUSEOUT should be 8 and is.
PASS: window.Event.MOUSEMOVE should be 16 and is.
PASS: window.Event.MOUSEDRAG should be 32 and is.
PASS: window.Event.CLICK should be 64 and is.
PASS: window.Event.DBLCLICK should be 128 and is.
PASS: window.Event.KEYDOWN should be 256 and is.
PASS: window.Event.KEYUP should be 512 and is.
PASS: window.Event.KEYPRESS should be 1024 and is.
PASS: window.Event.DRAGDROP should be 2048 and is.
PASS: window.Event.FOCUS should be 4096 and is.
PASS: window.Event.BLUR should be 8192 and is.
PASS: window.Event.SELECT should be 16384 and is.
PASS: window.Event.CHANGE should be 32768 and is.
